requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
bleach==6.2.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-filter==25.2
django-summernote==0.8.20.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
